backend/requirements.txt
fastapi
uvicorn
motor
redis
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
email-validator
langgraph
langchain-core
langchain-google-genai==4.2.0
amadeus
